+++ smokeping/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Version bump for new quarterlies
 * Add parameter to enable alternate networking
+* Fix smokeping config file issues, create directories second time
 
 ---